Sync logic for docker components
