+++ pkg/devfile/adapters/docker/component/adapter.go
@@ -3,6 +3,7 @@ package component
 import (
 	"fmt"
 
+	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/mount"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
@@ -11,8 +12,8 @@ import (
 	adaptersCommon "github.com/openshift/odo/pkg/devfile/adapters/common"
 	"github.com/openshift/odo/pkg/devfile/adapters/docker/storage"
 	"github.com/openshift/odo/pkg/devfile/adapters/docker/utils"
-	"github.com/openshift/odo/pkg/exec"
 	"github.com/openshift/odo/pkg/lclient"
+	"github.com/openshift/odo/pkg/sync"
 )
 
 // New instantiantes a component adapter
@@ -54,13 +55,29 @@ func (a Adapter) Push(parameters common.PushParameters) (err error) {
 		return errors.Wrap(err, "unable to create or update component")
 	}
 
-	cmd := []string{"/bin/sh", "-c", "./loop.sh"}
-	glog.V(3).Infof("MJF hola %v", cmd)
-	// err = a.Client.ExecCMDInContainer("", "ac758bf7fb60", cmd, nil, nil, nil, false)
-	err = exec.ExecuteCommand(&a.Client, "", "ac758bf7fb60", cmd, parameters.Show)
+	containers := utils.GetComponentContainers(a.Client, a.ComponentName)
+	// Find at least one pod with the source volume mounted, error out if none can be found
+	containerID, err := getFirstContainerWithSourceVolume(containers)
 	if err != nil {
-		return errors.Wrap(err, "unable to exec component")
+		return errors.Wrapf(err, "error while retrieving container for component: %s", a.ComponentName)
 	}
+
+	// Get a sync adapter. Check if project files have changed and sync accordingly
+	syncAdapter := sync.New(a.AdapterContext, &a.Client)
+	// podName is set to empty string on docker
+	// podChanged is set to false, since docker volume is always present even if container goes down
+	err = syncAdapter.CheckProjectFiles(parameters, "", containerID, false, componentExists)
+	if err != nil {
+		return errors.Wrapf(err, "Failed to sync to component with name %s", a.ComponentName)
+	}
+
+	// cmd := []string{"/bin/sh", "-c", "/tmp/loop.sh"}
+	// glog.V(3).Infof("MJF hola %v", cmd)
+	// // err = a.Client.ExecCMDInContainer("", "ac758bf7fb60", cmd, nil, nil, nil, false)
+	// err = exec.ExecuteCommand(&a.Client, "", "e91892812d10", cmd, parameters.Show)
+	// if err != nil {
+	// 	return errors.Wrap(err, "unable to exec component")
+	// }
 
 	return nil
 }
@@ -68,6 +85,22 @@ func (a Adapter) Push(parameters common.PushParameters) (err error) {
 // DoesComponentExist returns true if a component with the specified name exists, false otherwise
 func (a Adapter) DoesComponentExist(cmpName string) bool {
 	return utils.ComponentExists(a.Client, cmpName)
+}
+
+// getFirstContainerWithSourceVolume returns the first container that set mountSources: true
+// Because the source volume is shared across all components that need it, we only need to sync once,
+// so we only need to find one container. If no container was found, that means there's no
+// container to sync to, so return an error
+func getFirstContainerWithSourceVolume(containers []types.Container) (string, error) {
+	for _, c := range containers {
+		for _, mount := range c.Mounts {
+			if mount.Destination == lclient.OdoSourceVolumeMount {
+				return c.ID, nil
+			}
+		}
+	}
+
+	return "", fmt.Errorf("In order to sync files, odo requires at least one component in a devfile to set 'mountSources: true'")
 }
 
 // Delete attempts to delete the component with the specified labels, returning an error if it fails
